Changing Lamport so that one can store the secret seed instead of whole private key.
This change will allow for far smaller Merkle trees, where Lamport Keypairs
are generated on-the-fly when needed for tree generation or usage, but are
stored only as a 2048-byte secret seed when not in use. Compared to the
previous size of a private key alone (524.3 kilobytes!), this is a huge
improvement.